Refactor SourceCodeReader interface and add tests
We're moving from return null values to return java Optional wrapper, so the new SourceCodeReader interface adapts to this.
It also provides a new method for peek characters before consuming them.

In regard with the DefaultSourceCodeReader.java implementation, it's been renamed to BufferedSourceCodeReader, and follows DI principle by receiving a BufferedReader instead of creating one itself.
This change allows us to make better testing and to not depend of actual files to test the class.