Add sequence package
This package establishes what is a Sequence and what is a MarkableSequence. This interfaces will be use to refactor SourceCodeReader, converting it in a MarkableSequence, so clients can go back after reading characters.

This interfaces and specially the implementation ListMarkableSequence, may be useful for future stages, since it provides a way of reading sequentially buy going back if needed