Redefine LexicalNode and NodeBranch interfaces
The LexicalNode interface is now more generic, providing a method to process characters and return an element of any type. Also they now don't have different types of branches, there's only one addBranch method. The nodes will simply iterate through their branches and returning the value of the first branch that could successfully delegate to another node. If no branch could delegate, then the node will have to return an element on its own or throw a LexicalException.

A NodeBranch will simply have a filter and a next LexicalNode, given a certain character that passes the filter, the branch will delegate the processing of characters to the next node and return the value returned by the node. However, this is the basic implementation of a NodeBranch, since you could decorate it to perform extra logic (like Try-Branches) or altering the result of the node's returned element.