Re-implement automata package
With the new LexicalNode and NodeBranch interfaces, the automata package have been re-implemented.
The main class is DefaultLexicalNode, which has the basic iterate-delegate behavior, but when is not able to delegate, has a LexicalNodeStrategy to decide what to do.

The implemented strategies allows us to return a Token, throw a LexicalException, or don't do anything and return null.

Implementations of NodeBranch have been added too, being the DefaultNodeBranch class the one which implements basic behavior and the other ones allow clients to decorate the former with extra behaviour.

Lastly, two node builders have been created, these configure a DefaultLexicalNode so that it represent either a TokenizerNode or an OmitterNode.