Add MiniJavaLexicalAnalyzer
This analyzer uses all the automata package to generate the Tokens of a source code file.
A helper class has been created to simplify the code of tokenizing initial state, this class will skip unwanted characters in a loop, and then try to generate the next token.
So, the MiniJavaLexicalAnalyzer is just a set of static nodes that together form the lexical automata of the language.