Unify predefined symbols classes with user-created symbols
All this classes now belongs to the same hierarchy tree, but most of the methods of ClassSymbol were moved to UserClassSymbol, and the (now) interface ClassSymbol kept the most higher level ones.

This modification implies that now every UserClassSymbol extends from a default class that is a PredefinedClass given by the SymbolTable.