Fix allMethods overwrite bug
We were using a HashSetMultimap to generate the map with all the methods of a class.
But since we were first adding the inherited ones, the methods that overwrite those and thus were
considered equal by the set, were discarded. The fix is to reverse the order, adding first the methods
of the class and then the inherited ones.
This is the same for attributes in the case of UserClassSymbol